python-service/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain_huggingface